# libs/api-types/schema/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HomeAssist API",
    "version": "1.0.0",
    "description": "API for HomeAssist email processing and management system"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "required": ["statusCode", "message"]
      },
...
      "ProcessingType": {
        "type": "string",
        "enum": ["DATE_RANGE", "RECURRING", "SPECIFIC_DATES"]
      },
      "LlmFocus": {
        "type": "string",
        "enum": ["general", "sentiment", "urgency"]
      },
      "CreateProcessingScheduleDto": {
        "type": "object",
        "properties": {
          "userId": { "type": "string", "format": "uuid" },
          "emailAccountId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "processingType": { "$ref": "#/components/schemas/ProcessingType" },
          "dateRangeFrom": { "type": "string", "format": "date-time" },
          "dateRangeTo": { "type": "string", "format": "date-time" },
          "cronExpression": { "type": "string" },
          "timezone": { "type": "string" },
          "specificDates": { "type": "array", "items": { "type": "string", "format": "date-time" } },
          "batchSize": { "type": "integer", "minimum": 1, "maximum": 20 },
          "emailTypePriorities": { "type": "object", "additionalProperties": { "type": "string" } },
          "senderPriorities": { "type": "object", "additionalProperties": { "type": "string" } },
          "llmFocus": { "$ref": "#/components/schemas/LlmFocus" },
          "isEnabled": { "type": "boolean" },
          "isDefault": { "type": "boolean" }
        },
        "required": ["userId", "emailAccountId", "name", "processingType"]
      },
      "UpdateProcessingScheduleDto": {
        "type": "object",
        "properties": {
          "userId": { "type": "string", "format": "uuid" },
          "emailAccountId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "processingType": { "$ref": "#/components/schemas/ProcessingType" },
          "dateRangeFrom": { "type": "string", "format": "date-time" },
          "dateRangeTo": { "type": "string", "format": "date-time" },
          "cronExpression": { "type": "string" },
          "timezone": { "type": "string" },
          "specificDates": { "type": "array", "items": { "type": "string", "format": "date-time" } },
          "batchSize": { "type": "integer", "minimum": 1, "maximum": 20 },
          "emailTypePriorities": { "type": "object", "additionalProperties": { "type": "string" } },
          "senderPriorities": { "type": "object", "additionalProperties": { "type": "string" } },
          "llmFocus": { "$ref": "#/components/schemas/LlmFocus" },
          "isEnabled": { "type": "boolean" },
          "isDefault": { "type": "boolean" }
        }
      },
      "ProcessingSchedule": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "emailAccountId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "isEnabled": { "type": "boolean" },
          "isDefault": { "type": "boolean" },
          "processingType": { "$ref": "#/components/schemas/ProcessingType" },
          "dateRangeFrom": { "type": "string", "format": "date-time", "nullable": true },
          "dateRangeTo": { "type": "string", "format": "date-time", "nullable": true },
          "cronExpression": { "type": "string", "nullable": true },
          "timezone": { "type": "string" },
          "specificDates": { "type": "array", "items": { "type": "string", "format": "date-time" }, "nullable": true },
          "batchSize": { "type": "integer" },
          "emailTypePriorities": { "type": "object", "additionalProperties": { "type": "string" } },
          "senderPriorities": { "type": "object", "additionalProperties": { "type": "string" } },
          "llmFocus": { "$ref": "#/components/schemas/LlmFocus" },
          "lastExecutedAt": { "type": "string", "format": "date-time", "nullable": true },
          "nextExecutionAt": { "type": "string", "format": "date-time", "nullable": true },
          "totalExecutions": { "type": "integer" },
          "successfulExecutions": { "type": "integer" },
          "failedExecutions": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "userId", "emailAccountId", "name", "isEnabled", "isDefault", "processingType", "timezone", "batchSize", "emailTypePriorities", "senderPriorities", "llmFocus", "totalExecutions", "successfulExecutions", "failedExecutions", "createdAt", "updatedAt"]
      },
      "ProcessingScheduleWithAccount": {
        "type": "object",
        "allOf": [
          { "$ref": "#/components/schemas/ProcessingSchedule" }
        ],
        "properties": {
          "emailAccount": {
            "type": "object",
            "properties": {
              "email": { "type": "string", "format": "email" },
              "displayName": { "type": "string" }
            },
            "required": ["email", "displayName"]
          }
        }
      },
      "ValidationResult": {
        "type": "object",
        "properties": {
          "valid": { "type": "boolean" },
          "errors": { "type": "array", "items": { "type": "string" } },
          "warnings": { "type": "array", "items": { "type": "string" } },
          "cronConflicts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "conflictTime": { "type": "string", "format": "date-time" },
                "conflictingSchedules": { "type": "array", "items": { "type": "string" } },
                "suggestedAlternatives": { "type": "array", "items": { "type": "string", "format": "date-time" } }
              }
            }
          }
        },
        "required": ["valid", "errors", "warnings"]
      },
      "ScheduleExecutionStatus": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "scheduleId": { "type": "string", "format": "uuid" },
          "scheduleName": { "type": "string" },
          "status": { "type": "string", "enum": ["RUNNING", "COMPLETED", "FAILED", "CANCELLED", "PENDING"] },
          "progress": {
            "type": "object",
            "properties": {
              "totalBatches": { "type": "integer" },
              "completedBatches": { "type": "integer" },
              "totalEmails": { "type": "integer" },
              "processedEmails": { "type": "integer" },
              "failedEmails": { "type": "integer" },
              "completionPercentage": { "type": "number" }
            },
            "required": ["totalBatches", "completedBatches", "totalEmails", "processedEmails", "failedEmails", "completionPercentage"]
          },
          "timing": {
            "type": "object",
            "properties": {
              "startedAt": { "type": "string", "format": "date-time" },
              "completedAt": { "type": "string", "format": "date-time" },
              "estimatedCompletion": { "type": "string", "format": "date-time" },
              "processingDuration": { "type": "number" }
            },
            "required": ["startedAt"]
          },
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "details": { "type": "object" }
            }
          }
        },
        "required": ["id", "scheduleId", "scheduleName", "status", "progress", "timing"]
      },
...
      "ProcessingAnalytics": {
        "type": "object",
        "properties": {
          "userId": { "type": "string", "format": "uuid" },
          "totalSchedules": { "type": "integer" },
          "activeSchedules": { "type": "integer" },
          "totalExecutions": { "type": "integer" },
          "successfulExecutions": { "type": "integer" },
          "failedExecutions": { "type": "integer" },
          "averageProcessingTime": { "type": "number" },
          "emailsProcessedToday": { "type": "integer" },
          "emailsProcessedThisWeek": { "type": "integer" },
          "emailsProcessedThisMonth": { "type": "integer" },
          "recentExecutions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "scheduleName": { "type": "string" },
                "status": { "type": "string" },
                "startedAt": { "type": "string", "format": "date-time" },
                "completedAt": { "type": "string", "format": "date-time" },
                "processedEmails": { "type": "integer" },
                "failedEmails": { "type": "integer" }
              },
              "required": ["id", "scheduleName", "status", "startedAt", "processedEmails", "failedEmails"]
            }
          }
        },
        "required": ["userId", "totalSchedules", "activeSchedules", "totalExecutions", "successfulExecutions", "failedExecutions", "averageProcessingTime", "emailsProcessedToday", "emailsProcessedThisWeek", "emailsProcessedThisMonth", "recentExecutions"]
      },
      "BulkUpdateResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "updatedCount": { "type": "integer" },
          "errors": { "type": "array", "items": { "type": "string" } }
        },
        "required": ["success", "updatedCount"]
      },
      "ScheduleDetails": {
        "type": "object",
        "properties": {
          "schedule": { "$ref": "#/components/schemas/ProcessingSchedule" },
          "executionStats": {
            "type": "object",
            "properties": {
              "totalExecutions": { "type": "integer" },
              "successfulExecutions": { "type": "integer" },
              "failedExecutions": { "type": "integer" },
              "averageProcessingTime": { "type": "number" },
              "lastExecutionAt": { "type": "string", "format": "date-time" }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/processing-schedules": {
      "get": {
        "tags": ["Schedules"],
        "summary": "Get user processing schedules",
        "parameters": [
          { "name": "userId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "List of schedules",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProcessingScheduleWithAccount" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Schedules"],
        "summary": "Create processing schedule",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProcessingScheduleDto" } } } },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProcessingSchedule" } } } } }
      }
    },
    "/processing-schedules/{id}": {
      "put": {
        "tags": ["Schedules"],
        "summary": "Update processing schedule",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProcessingScheduleDto" } } } },
        "responses": { "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProcessingSchedule" } } } } }
      },
      "delete": {
        "tags": ["Schedules"],
        "summary": "Delete processing schedule",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "responses": { "200": { "description": "Deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } } }
      }
    },
    "/processing-schedules/{id}/execute": {
      "post": {
        "tags": ["Schedules"],
        "summary": "Execute schedule manually",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "responses": { "200": { "description": "Execution started", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "executionId": { "type": "string", "format": "uuid" } } } } } } }
      }
    },
    "/processing-schedules/{id}/status": {
      "get": {
        "tags": ["Schedules"],
        "summary": "Get schedule execution status",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "responses": { "200": { "description": "Status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleExecutionStatus" } } } } }
      }
    },
    "/processing-schedules/validate": {
      "post": {
        "tags": ["Schedules"],
        "summary": "Validate schedule configuration",
        "parameters": [ { "name": "excludeId", "in": "query", "schema": { "type": "string", "format": "uuid" } } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProcessingScheduleDto" } } } },
        "responses": { "200": { "description": "Validation result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationResult" } } } } }
      }
    },
    "/processing-schedules/check-conflicts": {
      "post": {
        "tags": ["Schedules"],
        "summary": "Check schedule conflicts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cronExpression": { "type": "string" },
                  "timezone": { "type": "string" },
                  "specificDates": { "type": "array", "items": { "type": "string", "format": "date-time" } },
                  "excludeId": { "type": "string", "format": "uuid" }
                },
                "required": ["timezone"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conflict check result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hasConflicts": { "type": "boolean" },
                    "conflicts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "conflictTime": { "type": "string", "format": "date-time" },
                          "conflictingSchedules": { "type": "array", "items": { "type": "string" } },
                          "suggestedAlternatives": { "type": "array", "items": { "type": "string", "format": "date-time" } }
                        }
                      }
                    }
                  },
                  "required": ["hasConflicts", "conflicts"]
                }
              }
            }
          }
        }
      }
    },
...
    "/processing-schedules/analytics/{userId}": {
      "get": {
        "tags": ["Schedules"],
        "summary": "Get processing analytics",
        "parameters": [ { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "responses": { "200": { "description": "Analytics", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProcessingAnalytics" } } } } }
      }
    },
    "/processing-schedules/bulk-enable": {
      "post": {
        "tags": ["Schedules"],
        "summary": "Bulk enable schedules",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "scheduleIds": { "type": "array", "items": { "type": "string", "format": "uuid" } } }, "required": ["scheduleIds"] } } } },
        "responses": { "200": { "description": "Bulk enable result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BulkUpdateResponse" } } } } }
      }
    },
    "/processing-schedules/bulk-disable": {
      "post": {
        "tags": ["Schedules"],
        "summary": "Bulk disable schedules",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "scheduleIds": { "type": "array", "items": { "type": "string", "format": "uuid" } } }, "required": ["scheduleIds"] } } } },
        "responses": { "200": { "description": "Bulk disable result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BulkUpdateResponse" } } } } }
      }
    },
    "/processing-schedules/{id}/details": {
      "get": {
        "tags": ["Schedules"],
        "summary": "Get schedule details",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "responses": { "200": { "description": "Details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleDetails" } } } }, "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      }
    }
  },
...
}